week-1/Fast API App/requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
jinja2
pwdlib[argon2]
python-jose
python-multipart